requirements.txt
gym~=0.26.2
gymnasium~=0.29.1
imgui~=2.0.0
matplotlib~=3.8.0
networkx~=3.0
numpy~=1.24.1
osmnx~=1.9.1
Pillow~=10.0.1
pygame~=2.5.2
scipy~=1.11.3
Shapely~=2.0.3
torch~=2.0.1+cu117
ezdxf~=1.1.4
geopandas~=0.14.3
numba
scikit-learn~=1.4.1.post1
pandas~=2.2.1
tqdm~=4.66.1
pympler~=1.0.1
opencv-python~=4.8.0.76
pyautogui~=0.9.54
moderngl~=5.10.0
PyQt5~=5.15.10
moderngl_window
swifter
geovoronoi~=0.4.0
pythonnet
PyOpenGL
torchsummary
pydantic~=2.6.2
pyrr~=0.10.3